requirements.txt
Babel >= 2.7.0
Image >= 1.5.27
imageio >= 2.5.0
matplotlib >= 3.1.1
numpy >= 1.17.0
psutil >= 5.6.3
pytz >= 2019.1
pyyaml >= 5.1.1
tkfilebrowser >= 2.3.1
pywin32
